requirements.txt
Django==1.7.5
Pillow==2.7.0
dj-database-url==0.3.0
django-braces==1.4.0
django-crispy-forms==1.4.0
django-endless-pagination==2.0
ipython==3.0.0
psycopg2==2.6
six==1.9.0
